raise full errors if there is a bad response from the API
when there is an unexpected result from the API, this gem should be able
to communicate that to the consumer. This is already done for a couple
of `get` requests, so why not handle it for any error type?

these errors are all grouped underneath Metrc::Errors::RequestError so
that they can be rescued from as a group.